fix jf17 filter

+++ reconstruction/CaloRec/src/CaloNtupleMaker.cxx
@@ -590,7 +590,7 @@ void CaloNtupleMaker::Fill( EventContext &ctx , TTree *tree, xAOD::seed_t seed, 
           mc_cl_cell_energy->push_back( cell->energy() );
           // Get all necessary ids to locate the cell outside
           mc_cl_cell_channel_eta->push_back( cell->parent()->channelEta() );
-          mc_cl_cell_channel_phi->push_back( (int)cell->parent()->channelPhi() );
+          mc_cl_cell_channel_phi->push_back( cell->parent()->channelPhi() );
           mc_cl_cell_sampling->push_back( (int)cell->sampling() );
         }
       }
@@ -649,7 +649,7 @@ void CaloNtupleMaker::Fill( EventContext &ctx , TTree *tree, xAOD::seed_t seed, 
           cl_cell_energy->push_back( cell->energy() );
           // Get all necessary ids to locate the cell outside
           cl_cell_channel_eta->push_back( cell->parent()->channelEta() );
-          cl_cell_channel_phi->push_back( (int)cell->parent()->channelPhi() );
+          cl_cell_channel_phi->push_back( cell->parent()->channelPhi() );
           cl_cell_sampling->push_back( (int)cell->sampling() );
